requirements.txt
pyTelegramBotAPI==4.23.0
google-genai==1.68.0
ddgs==9.11.4
schedule==1.2.2
requests==2.32.3
